+++ resources/release-notes.md
@@ -1,3 +1,3 @@
-## Package
+## App
 
-- Add utility method for determining installed AddOn version
+- (All) `datetime` slider labels do not display seconds when time range is > 6 hours